Refactoring so that an EventStream is always returned.
Not sure if this is a good idea, but the rational behind it is that it's possible to _create_ an aggregate the same way you would update it. I.e. you simply load the event stream, regardless if events exists for the stream id or not, call the domain model and append events with the current version (0 in case no events exists).